+++ src/DetailsView/components/cards/instance-details.tsx
