+++ src/DetailsView/components/cards/rule-resources.tsx
