+++ src/DetailsView/components/cards/rules-with-instances.scss
