+++ src/DetailsView/components/details-list-issues-view.tsx
@@ -13,7 +13,7 @@ import { VisualizationScanResultData } from '../../common/types/store-data/visua
 import { VisualizationStoreData } from '../../common/types/store-data/visualization-store-data';
 import { VisualizationType } from '../../common/types/visualization-type';
 import { DetailsViewToggleClickHandlerFactory } from '../handlers/details-view-toggle-click-handler-factory';
-import { UnifiedStatusResults } from './cards/failed-instances-section-v2';
+import { UnifiedStatusResults } from './cards/failed-instances-section';
 import { IssuesTable, IssuesTableDeps } from './issues-table';
 import { IssuesTableHandler } from './issues-table-handler';
