+++ src/tests/unit/tests/DetailsView/components/cards/rule-content.test.tsx
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 import { shallow } from 'enzyme';
 import * as React from 'react';
-import { RuleContent, RuleContentProps } from '../../../../../../DetailsView/components/cards/rule-content-v2';
+import { RuleContent, RuleContentProps } from '../../../../../../DetailsView/components/cards/rule-content';
 
 describe('RuleContent', () => {
     it('renders', () => {
