+++ src/tests/unit/tests/DetailsView/components/cards/rules-with-instances.test.tsx
@@ -7,7 +7,7 @@ import { IMock, Mock } from 'typemoq';
 
 import { NamedFC, ReactFCWithDisplayName } from '../../../../../../common/react/named-fc';
 import { CollapsibleComponentCardsProps } from '../../../../../../DetailsView/components/cards/collapsible-component-cards';
-import { RulesWithInstances, RulesWithInstancesDeps } from '../../../../../../DetailsView/components/cards/rules-with-instances-v2';
+import { RulesWithInstances, RulesWithInstancesDeps } from '../../../../../../DetailsView/components/cards/rules-with-instances';
 import { exampleUnifiedRuleResult } from './sample-view-model-data';
 
 describe('RulesWithInstances', () => {
