+++ src/tests/unit/tests/reports/components/report-sections/report-body.test.tsx
@@ -8,7 +8,7 @@ import { ReportBody, ReportBodyProps } from 'reports/components/report-sections/
 import { ReportSectionFactory, SectionProps } from 'reports/components/report-sections/report-section-factory';
 import { Mock } from 'typemoq';
 
-import { FailedInstancesSectionDeps } from '../../../../../../DetailsView/components/cards/failed-instances-section-v2';
+import { FailedInstancesSectionDeps } from '../../../../../../DetailsView/components/cards/failed-instances-section';
 import { exampleUnifiedStatusResults } from '../../../DetailsView/components/cards/sample-view-model-data';
 
 describe('ReportBody', () => {
